+++ src/api/objects/toot.ts
@@ -822,7 +822,7 @@ export default class Toot implements TootObj {
     // Fetch all the data we need to set dependent properties and set them on the toots.
     // If 'source' arg is proivded we set it as the Toot.source prop and avoid doing an isDeepInspect completion
     static async completeToots(toots: TootLike[], logger: Logger, source?: string): Promise<Toot[]> {
-        logger = logger.tempLogger(`completeToots("${source || 'NO_SOURCE'}")`);
+        logger = logger.tempLogger(`completeToots(${source || ''})`);
         const isDeepInspect = !source;
         const startedAt = new Date();